[RPC] Use leaky singleton for current RPC agent global
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41591

This allows to delay its construction until it's first used and, therefore, ensure its destruction before all global static objects. In particular, if we didn't do this, and if the user forgot to explicitly call `rpc.shutdown()`, then the static destruction order could first destroy the glog mutexes and only then destroy the global static shared_ptr holding the RPC agent, which would only shut down at that moment, and if it were to encounter an error and attempt to log it would fail and crash the process.
ghstack-source-id: 108015839

Differential Revision: [D22597100](https://our.internmc.facebook.com/intern/diff/D22597100/)