Call qnnpack's conv setup only if input pointer has changed.
Summary:
With caching allocator we have increased the likelihood of getting the
same input pointer. With that we can cache qnnpack operator and input
pointer and check if the input pointer is the same. If so we can skip
setup step.

Test Plan:
Ran one of the quantized models to observe
1. No pagefaults due to indirection buffer reallocation.
2. Much less time spent in indirection buffer population.

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 2f1adafd20cffc4c1222eef0c0849f04a4a596b2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42008