[TensorExpr] Implement shape inference for TE.
Since TE operates on a limited subset of ops with a well-defined
semantics, we can easily infer shapes of intermediate and output tensors
given shapes of the inputs.

There is a couple of ops that are not yet supported in the shape
inference, once we add them we could relax the shape info requirements
in the TE fuser: currently it requires all values in the fusion group to
have shapes known and we can change it to only inputs.

ghstack-source-id: fe6c00c5b79d85237e5082f7dcf4dd2781044975
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41451