ecr_gc: Iterate through all tags, reduce prints
There's a potential for multiple tags to be created for the same digest
so we should iterate through all potential tags so that we're not
deleting digests that are associated with tags that we actually want.

Also, reduced the number of prints in this script to only the absolutely
necessary prints.

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>

ghstack-source-id: 5a48e0d84600ac3051657979815c8acc85613960
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42492