Workaround max(..., out=) requiring same strides in adamax.
Adamax calls max(input, 0, keepdim=False, out=(input, input.new().long()).
But the implementation of the out variant requires that the two tensors have the same strides.

Perhaps we should fix the underlying problem, but the use case is pretty esoteric, so for now, we just force the strides to be the same.

ghstack-source-id: 421ea311da92124d3f87c9588ee699680c899431
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42543