[JIT] Add property support to TorchScript classes
**Summary**
This commit adds support for properties to TorchScript classes,
specifically for getters and setters. They are implemented essentially
as pointers to the methods that the corresponding decorators decorate,
which are treated like regular class methods. Deleters for properties
are considered to be out of scope (and probably useless for TorchScript
anyway).

**Test Plan**
This commit adds a unit test for a class with a property that has both
getter and setter and one that has only a getter.

`python test/test_jit.py TestClassType.test_properties`

ghstack-source-id: 0a5c84d46284b2c7b71d1b13f06eb636f52ba250
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42389